requirements.txt
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=0.24.0
yfinance>=0.1.70
beautifulsoup4>=4.9.0
requests>=2.25.0
plotly>=5.3.0
python-dotenv>=0.19.0
nltk>=3.6.0
pandas-datareader>=0.10.0
scipy>=1.7.0
pypfopt>=1.5.0 